Migração do MainActivity para EvadaoScreen Otimização do MainActivity
